Merge in ECOMR4/ecomr4-list-testing from R4LA-432-Create-Group-UI to master

* commit '1746fad0b2586f393cdd95573a80b1b4008ca4e8': (33 commits)
  Fix pricing api Added email for panamax channel and list channel
  New docker image
  Smaller update
  Refactors and fixes
  Added into cypress/support/index  the testing-library
  Added new validation to check ellipsis on the banner image
  These features won't run in the regression until data issues is resolved
  Added step definition and page for new test Create Group UI
  Refactor features with new user data
  -Updated selectors -Updated email to
  Added final report-email config
  new test5
  new test4
  new test3
  new test2
  new test1
  new test
  Added after mocha
  Fixed email functionality
  Updated path to delete folder on  delete-results.js
  ...

+++ cypress/reports/setup/delete-results.js
@@ -1,9 +1,18 @@
 #!/usr/bin/env node
+//     #!/bin/bash
+
+
 const rimraf = require('rimraf')
 const chalk = require('chalk')
+const fs = require("fs");
 
-const testResultsDir = '../results'
+const testResultsDir = '../results/'
 
-rimraf(testResultsDir, () => {
-  console.info(chalk.yellow(`    Deleted former test results on : ${testResultsDir}`))
-})
+if(fs.existsSync(testResultsDir)) {
+  rimraf(testResultsDir, () => {
+    console.info(chalk.yellow(`    Deleted former test results on : ${testResultsDir}\n`))
+  });
+} else {
+  console.info(chalk.yellow(`    Result folder has not been created\n`))
+}
+
